Kubernetes go mod conquered - now we can use utils from k8s. Yeah
