+++ go.mod
@@ -5,31 +5,29 @@ go 1.13
 require (
 	github.com/Masterminds/semver v1.4.2 // indirect
 	github.com/VividCortex/gohistogram v1.0.0 // indirect
-	github.com/aws/aws-sdk-go v1.16.26 // indirect
+	github.com/aws/aws-sdk-go v1.30.7 // indirect
 	github.com/cespare/xxhash/v2 v2.1.1 // indirect
 	github.com/codahale/hdrhistogram v0.0.0-20161010025455-3a0bb77429bd // indirect
 	github.com/docker/go-metrics v0.0.0-20180209012529-399ea8c73916 // indirect
 	github.com/docker/libtrust v0.0.0-20160708172513-aabc10ec26b7 // indirect
 	github.com/google/uuid v1.1.1
 	github.com/googleapis/gnostic v0.3.1 // indirect
-	github.com/gorilla/context v1.1.1 // indirect
-	github.com/gorilla/mux v1.6.2 // indirect
+	github.com/gorilla/mux v1.7.4 // indirect
 	github.com/gorilla/websocket v1.4.1
 	github.com/lunarway/color v1.7.0
 	github.com/mattn/go-colorable v0.1.4 // indirect
 	github.com/mattn/go-isatty v0.0.10 // indirect
 	github.com/nlopes/slack v0.6.0
-	github.com/opencontainers/image-spec v1.0.1 // indirect
 	github.com/opentracing/opentracing-go v1.1.0
 	github.com/pkg/errors v0.9.1
 	github.com/pkg/term v0.0.0-20180730021639-bffc007b7fd5 // indirect
 	github.com/prometheus/client_golang v1.2.1
 	github.com/prometheus/procfs v0.0.8 // indirect
-	github.com/ryanuber/go-glob v0.0.0-20160226084822-572520ed46db // indirect
+	github.com/ryanuber/go-glob v1.0.0 // indirect
 	github.com/spf13/cobra v0.0.5
 	github.com/spf13/pflag v1.0.5
 	github.com/streadway/amqp v0.0.0-20200108173154-1c71cc93ed71
-	github.com/stretchr/testify v1.4.0
+	github.com/stretchr/testify v1.5.1
 	github.com/uber/jaeger-client-go v2.20.1+incompatible
 	github.com/uber/jaeger-lib v2.2.0+incompatible
 	github.com/weaveworks/flux v0.0.0-20190411145155-fea56bc3feee
@@ -38,7 +36,6 @@ require (
 	go.uber.org/multierr v1.4.0
 	go.uber.org/zap v1.13.0
 	golang.org/x/lint v0.0.0-20191125180803-fdd1cda4f05f // indirect
-	golang.org/x/net v0.0.0-20191126235420-ef20fe5d7933 // indirect
 	golang.org/x/oauth2 v0.0.0-20191202225959-858c2ad4c8b6 // indirect
 	golang.org/x/sys v0.0.0-20191128015809-6d18c012aee9 // indirect
 	golang.org/x/time v0.0.0-20191024005414-555d28b269f0 // indirect
@@ -49,8 +46,55 @@ require (
 	gopkg.in/src-d/go-git.v4 v4.13.1
 	gopkg.in/yaml.v2 v2.2.8
 	k8s.io/api v0.18.0
+	k8s.io/apiextensions-apiserver v0.0.0
 	k8s.io/apimachinery v0.18.0
+	k8s.io/apiserver v0.18.0
+	k8s.io/cli-runtime v0.18.0
 	k8s.io/client-go v0.18.0
-	k8s.io/kubectl v0.18.0
-	sigs.k8s.io/structured-merge-diff v0.0.0-20190525122527-15d366b2352e // indirect
+	k8s.io/cloud-provider v0.18.0
+	k8s.io/cluster-bootstrap v0.0.0
+	k8s.io/code-generator v0.18.0
+	k8s.io/component-base v0.18.0
+	k8s.io/cri-api v0.0.0
+	k8s.io/csi-translation-lib v0.18.0
+	k8s.io/kube-aggregator v0.0.0
+	k8s.io/kube-controller-manager v0.0.0
+	k8s.io/kube-proxy v0.0.0
+	k8s.io/kube-scheduler v0.0.0
+	k8s.io/kubectl v0.0.0
+	k8s.io/kubelet v0.0.0
+	k8s.io/kubernetes v1.18.0
+	k8s.io/legacy-cloud-providers v0.0.0
+	k8s.io/metrics v0.18.0
+	k8s.io/sample-apiserver v0.0.0
+)
+
+// k8s.io/kubernetes has a go.mod file that sets the version of the following
+// modules to v0.0.0. This causes go to throw an error. These need to be set
+// to a version for Go to process them. Here they are set to the same
+// revision as the marked version of Kubernetes. When Kubernetes is updated
+// these need to be updated as well.
+
+replace (
+	k8s.io/api => k8s.io/api v0.18.0
+	k8s.io/apiextensions-apiserver => k8s.io/apiextensions-apiserver v0.18.0
+	k8s.io/apimachinery => k8s.io/apimachinery v0.18.0
+	k8s.io/apiserver => k8s.io/apiserver v0.18.0
+	k8s.io/cli-runtime => k8s.io/cli-runtime v0.18.0
+	k8s.io/client-go => k8s.io/client-go v0.18.0
+	k8s.io/cloud-provider => k8s.io/cloud-provider v0.18.0
+	k8s.io/cluster-bootstrap => k8s.io/cluster-bootstrap v0.18.0
+	k8s.io/code-generator => k8s.io/code-generator v0.18.0
+	k8s.io/component-base => k8s.io/component-base v0.18.0
+	k8s.io/cri-api => k8s.io/cri-api v0.18.0
+	k8s.io/csi-translation-lib => k8s.io/csi-translation-lib v0.18.0
+	k8s.io/kube-aggregator => k8s.io/kube-aggregator v0.18.0
+	k8s.io/kube-controller-manager => k8s.io/kube-controller-manager v0.18.0
+	k8s.io/kube-proxy => k8s.io/kube-proxy v0.18.0
+	k8s.io/kube-scheduler => k8s.io/kube-scheduler v0.18.0
+	k8s.io/kubectl => k8s.io/kubectl v0.18.0
+	k8s.io/kubelet => k8s.io/kubelet v0.18.0
+	k8s.io/legacy-cloud-providers => k8s.io/legacy-cloud-providers v0.18.0
+	k8s.io/metrics => k8s.io/metrics v0.18.0
+	k8s.io/sample-apiserver => k8s.io/sample-apiserver v0.18.0
 )
