+++ internal/slack/slack.go
@@ -106,80 +106,6 @@ func (c *Client) PostSlackBuildStarted(email, title, titleLink, text, color stri
 	return respChannel, timestamp, err
 }
 
-//DEPRECATED
-func (c *Client) PostPrivateMessage(ctx context.Context, email string, podNotify *http.PodNotifyRequest) error {
-	if c.muteOptions.Kubernetes {
-		return nil
-	}
-	userID, err := c.getIdByEmail(ctx, email)
-	if err != nil {
-		return err
-	}
-	asUser := slack.MsgOptionAsUser(true)
-	switch podNotify.State {
-	case "CrashLoopBackOff":
-		_, _, err := c.client.PostMessageContext(ctx, userID, asUser, crashLoopBackOffErrorMessage(podNotify))
-		if err != nil {
-			return err
-		}
-	case "CreateContainerConfigError":
-		_, _, err := c.client.PostMessageContext(ctx, userID, asUser, createConfigErrorMessage(podNotify))
-		if err != nil {
-			return err
-		}
-	case "Running", "Ready":
-		_, _, err := c.client.PostMessageContext(ctx, userID, asUser, successMessage(podNotify))
-		if err != nil {
-			return err
-		}
-	default:
-		return errors.New("unknown pod state in post private message")
-	}
-	return nil
-}
-
-// DEPRECATED
-func successMessage(podNotify *http.PodNotifyRequest) slack.MsgOption {
-	return slack.MsgOptionAttachments(slack.Attachment{
-		Title:      fmt.Sprintf(":kubernetes: k8s (%s) :white_check_mark:", podNotify.Environment),
-		Text:       fmt.Sprintf("%s (%s)\nArtifact: *%s*", podNotify.Name, podNotify.State, podNotify.ArtifactID),
-		Color:      "#73bf69",
-		MarkdownIn: []string{"text", "fields"},
-	})
-}
-
-// DEPRECATED
-func createConfigErrorMessage(podNotify *http.PodNotifyRequest) slack.MsgOption {
-	messageField := slack.AttachmentField{
-		Title: "Error",
-		Value: fmt.Sprintf("```%s```", podNotify.Message),
-		Short: false,
-	}
-	return slack.MsgOptionAttachments(slack.Attachment{
-		Title:      fmt.Sprintf(":kubernetes: k8s (%s) :no_entry:", podNotify.Environment),
-		Text:       fmt.Sprintf("%s (*%s*)\nArtifact: *%s*", podNotify.Name, podNotify.State, podNotify.ArtifactID),
-		Color:      "#e24d42",
-		MarkdownIn: []string{"text", "fields"},
-		Fields:     []slack.AttachmentField{messageField},
-	})
-}
-
-// DEPRECATED
-func crashLoopBackOffErrorMessage(podNotify *http.PodNotifyRequest) slack.MsgOption {
-	logField := slack.AttachmentField{
-		Title: "Logs",
-		Value: fmt.Sprintf("```%s```", podNotify.Logs),
-		Short: false,
-	}
-	return slack.MsgOptionAttachments(slack.Attachment{
-		Title:      fmt.Sprintf(":kubernetes: k8s (%s) :no_entry:", podNotify.Environment),
-		Text:       fmt.Sprintf("%s (%s)\nArtifact: *%s*", podNotify.Name, podNotify.State, podNotify.ArtifactID),
-		Color:      "#e24d42",
-		MarkdownIn: []string{"text", "fields"},
-		Fields:     []slack.AttachmentField{logField},
-	})
-}
-
 type ReleaseOptions struct {
 	Service           string
 	ArtifactID        string
